requirements.txt
Flask==2.3.3
nltk==3.8.1
scikit-learn==1.3.0
numpy==1.24.3
tensorflow==2.13.0
python-dotenv==1.0.0
flask-cors==4.0.0
openai==0.27.8